requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
Django==4.2.4
ffmpeg-python==0.2.0
frozenlist==1.4.0
future==0.18.3
giphy-client==1.0.0
idna==3.4
imageio-ffmpeg==0.4.8
multidict==6.0.4
numpy==1.25.2
praw==7.7.1
prawcore==2.3.0
py-cord==2.4.1
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
update-checker==0.18.0
urllib3==2.0.4
websocket-client==1.6.2
yarl==1.9.2
